Add runtime timing

+++ src/Engine.Core/Transform.cs
@@ -2,14 +2,14 @@ namespace Engine.Core;
 
 public sealed class Transform : Component
 {
-    public float X { get; set; }
-    public float Y { get; set; }
+    public double X { get; set; }
+    public double Y { get; set; }
 
     public Transform(GameObject gameObject) : this(gameObject, 0f, 0f)
     {
         
     }
-    public Transform(GameObject gameObject, float x, float y) : base(gameObject)
+    public Transform(GameObject gameObject, double x, double y) : base(gameObject)
     {
         X = x;
         Y = y;
